+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.0.18
 - `Clean Flightplan`: new option to add/remove spaces after `//`. See "Clean Flightplan" → "Ajust Comments" setting.
+- New function `Switch FS9 ⟷ FSX`: changes the days of selected flightplans to either FSX or FS9 (increase/decrease day number).
 
 ### 1.0.17
 - `Create Aircraft` improvements: